Lagt til deltaY

+++ widget.h
@@ -20,7 +20,7 @@ class Widget {
 	Point _startpoint{};
 	int _height{0};
 	int _width{0};
-	int _deltaX{0};
+	float _deltaX{0};
 	int _deltaY{0};
 	int _angle{0};
 	bool _rotating{false};
@@ -32,7 +32,7 @@ class Widget {
 
 	public:
 	    Widget() { std::cout << "Widget()\n" ; }
-	    Widget(std::string fn, SDL_Renderer* rend, Point p={0, 0}, int w = 0, int h = 0, int dx = 0, int dy = 0, int a= 0 ): filename{fn}, 
+	    Widget(std::string fn, SDL_Renderer* rend, Point p={0, 0}, int w = 0, int h = 0, float dx = 0, int dy = 0, int a= 0 ): filename{fn}, 
                                                                                                                     //renderer{rend},
                                                                                                                     _startpoint{p},                                                                                                                    
                                                                                                                     _width{w},
@@ -100,7 +100,7 @@ class Widget {
 		int& height()  {
 		    return _height;
 		}
-		int& deltaX() {
+		float& deltaX() {
 		    return _deltaX;
 		}
 		int& deltaY() {
